isolate Result model tests
